requirements.txt
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
pydantic>=2.5.0
httpx>=0.26.0
pillow>=10.2.0
pytesseract>=0.3.10
python-multipart>=0.0.6
jinja2>=3.1.3
